restructured objects and functions
