+++ utils.py
@@ -14,3 +14,32 @@ manager_eff = {
     "A Nu Start" : 0.811,
     "ELC3" : 0.828
 }
+
+def extract_position(text) -> str:
+    '''function that extracts the position from Yahoo player html'''
+    try:
+        text = text.split(" - ")[1].split(" ")[0]
+    except:
+        pass
+    return text
+
+def extract_player_name(text) -> str:
+    '''function that extracts the player name from the Yahoo player html'''
+    text = str(text).split("-")[0].strip()
+    if text[-2:].upper() in team_abbrev:
+        player = text[:-2]
+    elif text[-3:].upper() in team_abbrev:
+        player = text[:-3]
+    else:
+        player = text
+    return player
+
+def extract_team(text) -> None | str:
+    '''function that extracts the team name from the Yahoo player html'''
+    text = str(text).split("-")[0].strip()
+    if text[-2:].upper() in team_abbrev:
+        return text[-2:].upper()
+    elif text[-3:].upper() in team_abbrev:
+        return text[-3:].upper()
+    else:
+        return None
